#3 Make plus icon appear when hovering over dropzone
