+++ frontend/src/pages/StartingConfig.tsx
@@ -1,6 +1,6 @@
 import { Grid, GridItem } from '@chakra-ui/react';
 import { FC } from 'react';
-import DropZone from '../components/dragdrop/DropTarget';
+import Dropzone from '../components/dragdrop/Dropzone';
 import { ItemTypes } from '../components/ItemTypeDefinitions';
 import SearchColumn from '../components/SearchColumn';
 import { Skills } from '../data/TestSkills';
@@ -10,7 +10,7 @@ const StartingConfig: FC = () => {
         <Grid templateColumns="repeat(3, 1fr)" gap={4}>
             <GridItem colSpan={2}>
                 <div>Starting configuration!</div>
-                <DropZone itemType={ItemTypes.SKILL} onItemDropped={console.log} />
+                <Dropzone itemType={ItemTypes.SKILL} onItemDropped={console.log} />
             </GridItem>
             <GridItem colSpan={1}>
                 <SearchColumn items={Skills} />
