+++ package.json
@@ -12,6 +12,10 @@
 		"@chakra-ui/react": "^2.2.1",
 		"@emotion/react": "^11.9.3",
 		"@emotion/styled": "^11.9.3",
+		"@fortawesome/fontawesome-svg-core": "^6.1.1",
+		"@fortawesome/free-regular-svg-icons": "^6.1.1",
+		"@fortawesome/free-solid-svg-icons": "^6.1.1",
+		"@fortawesome/react-fontawesome": "^0.2.0",
 		"@tauri-apps/api": "^1.0.1",
 		"framer-motion": "^6.3.16",
 		"react": "^18.0.0",
